+++ src/Artists/artists.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ArtistsController } from './artists.controller';
 import { ArtistsService } from './artists.service';
-import { TracksService } from '../Tracks/traks.service';
+import { TracksService } from '../Tracks/tracks.service';
 import { TracksModule } from '../Tracks/tracks.module';
 
 @Module({
